Major refactoring. Complete sensor interface overhaul. New AirT and replaced RelH sensor

+++ main.py
@@ -1,22 +1,18 @@
-import os
 import time
-import sqlite3
+import os
 import subprocess
 
 import RPi.GPIO as gpio
 
 import routines.config as config
 import routines.helpers as helpers
-import routines.queries as queries
+import routines.data as data
 
-proc_support = None
-proc_data = None
 
-# -- INIT GPIO AND LEDS --------------------------------------------------------
+# Set up and reset data and error indicator LEDs
 gpio.setwarnings(False)
 gpio.setmode(gpio.BCM)
 
-# Setup and reset the data and error LEDs
 gpio.setup(helpers.DATALEDPIN, gpio.OUT)
 gpio.output(helpers.DATALEDPIN, gpio.LOW)
 gpio.setup(helpers.ERRORLEDPIN, gpio.OUT)
@@ -29,51 +25,51 @@ time.sleep(2.5)
 gpio.output(helpers.DATALEDPIN, gpio.LOW)
 gpio.output(helpers.ERRORLEDPIN, gpio.LOW)
 
-# -- CHECK CONFIG --------------------------------------------------------------
-if config.load() == False: helpers.init_error(1)
 
-# -- CHECK INTERNAL DRIVE ------------------------------------------------------
+# Loading the configuration file
+if config.load() == False: helpers.init_error(0)
+
+# Check remaining space on internal computer drive
 free_space = helpers.remaining_space("/")
-if free_space == None or free_space < 1: helpers.init_error(2)
+if free_space == None or free_space < 1: helpers.init_error(1)
 
-# -- CHECK DATA DIRECTORY ------------------------------------------------------
+# Create data directory if it doesn't exist
 if not os.path.isdir(config.data_directory):
     try:
         os.makedirs(config.data_directory)
-    except: helpers.init_error(3)
+    except: helpers.init_error(2)
 
-# -- MAKE DATABASE -------------------------------------------------------------
+# Create database if it doesn't exist
 if not os.path.isfile(config.database_path):
     try:
-        with sqlite3.connect(config.database_path) as database:
-            cursor = database.cursor()
-            cursor.execute(queries.CREATE_REPORTS_TABLE)
-            cursor.execute(queries.CREATE_ENVREPORTS_TABLE)
-            cursor.execute(queries.CREATE_DAYSTATS_TABLE)
+        data.create_database(config.database_path)
+    except: helpers.init_error(3)
 
-            database.commit()
-    except: helpers.init_error(4)
-
-# -- CHECK CAMERA DRIVE --------------------------------------------------------
+# Check camera image storage drive is ready for use
 if config.camera_logging == True:
-    if not os.path.isdir(config.camera_directory): helpers.init_error(5)
-    if not os.path.ismount(config.camera_directory): helpers.init_error(6)
+    if not os.path.isdir(config.camera_directory): helpers.init_error(4)
+    if not os.path.ismount(config.camera_directory): helpers.init_error(5)
 
     free_space = helpers.remaining_space(config.camera_directory)
-    if free_space == None or free_space < 5: helpers.init_error(7)
+    if free_space == None or free_space < 5: helpers.init_error(6)
 
 
-# -- RUN SUPPORT SUBSYSTEM ----------------------------------------------------
+# Start support subsystem
 try:
-    proc_support = subprocess.Popen(["python3", "aws_support.py"])
-except: helpers.init_error(8)
+    with open(os.devnull, "w") as devnull:
+        proc_support = subprocess.Popen(["python3", "aws_support.py"],
+            stdout=devnull, stderr=devnull)
+except: helpers.init_error(7)
 
-# -- RUN DATA SUBSYSTEM -------------------------------------------------------
+# Start data subsystem
 try:
-    proc_data = subprocess.Popen(["python3", "aws_data.py"])
-    gpio.output(helpers.DATALEDPIN, gpio.HIGH)
-    gpio.output(helpers.ERRORLEDPIN, gpio.HIGH)
-    
+    with open(os.devnull, "w") as devnull:
+        subprocess.Popen(["python3", "aws_data.py"], stdout=devnull,
+            stderr=devnull)
+
 except:
-    proc_support.terminate()
-    helpers.init_error(9)
+    if proc_support.poll() == None: proc_support.terminate()
+    helpers.init_error(8)
+
+gpio.output(helpers.DATALEDPIN, gpio.HIGH)
+gpio.output(helpers.ERRORLEDPIN, gpio.HIGH)
